video_play_vs_comms.sh: first attempt at adding computation of drop rate
The drop rate is computed from the total number of frames of the video
and the number of frames dropped during the playback. Those data are
provided by the output of mplayer.
The computation is made using awk and therefore should add no new
dependencies to the benchmark suite. At the end of each test, stats
concerning the drop rate are added to the file containing the results.