comm_startup_lat.sh: use PID of interactive process to kill it
The comm_startup_lat.sh script uses the name of the interactive
process to kill it if the test lasts too long. Unfortunately, some
interactive processes don't preserve their name during their
execution, e.g. "oowriter" is just a wrapper script for a soffice.bin
executable. For this kind of processes, any attempt of killing them
using their initial name would fail.
This commit uses the PID of the bash which starts the interactive
process to kill it if the test lasts too long. For this purpose,
it was necessary to set that same bash as process group leader, so
that all its subprocesses can be recursively killed. This commit
also adds the removal of the "Stop-iterations" marker file and the
newly-introduced "current-pid" file as soon as the test begins, to
cleanup the test environment from remnants of previous executions.