run_all_tests_1.sh: change number of iterations and repetitions
Currently, the run_all_tests_1.sh script runs only 5 iterations for
all startup test scenarios involving oowriter. This makes the test
lighter, especially with write background workloads, that increase
the load of work imposed on the flusher thread(s) with the progress
of the single repetition.
This commit increases the number of iterations to 10 even for such
test scenarios; on the other hand, this commit decreases the number
of repetitions to 5.