First attempt at adding support for repetitions to video-playing tests
This commit attempts to add a first implementation for repetitions in
video playing tests. The generation of graphs uses as target quantity
the frame drop rate of tests. The label of the generated overall stats
is still the interactive command periodically executed during the
playback of the video, as it was before.