video_play_vs_comms.sh: use eval to invoke mplayer
One of the last commits concerning the video_play_vs_comms.sh script
changed the name of the video file used for video-playing tests. The
new filename contains spaces, and must be enclosed in quotes. Even if
it is, the "mplayer" command is kept in a variable: when the command
is executed, the quotes are trimmed out. This commit adds quoting to
the name of the file and uses "eval" to invoke mplayer.