comm_startup_lat.sh: improve robustness of the timeout mechanism
On a very slow disk, under heavy write background workloads, the
script might not be able to to write the current-pid file before
the killer process is run by the timeou and/or to write the
Stop-iterations file after a timeout is triggered. In both cases
the subsequent iterations of the test repetition may not stopped.
This commit attempts to make the mechanism more robust against
these adverse scenarios.

Signed-off-by: Paolo Valente <paolo.valente@unimore.it>
Tested-by: Alexander Spyridakis <a.spyridakis@virtualopensystems.com>
Signed-off-by: Alexander Spyridakis <a.spyridakis@virtualopensystems.com>