comm_startup_lat.sh: do not round down sleep time computed from raising time
Currently, when BFQ is selected as test scheduler, the comm_startup_lat.sh
script computes the initial sleep time from the raising time, made
available by BFQ in the raising_max_time tunable. However, the number of
seconds needed by the sleep command are rounded down; this might lead to
computing a sleep time which is almost a second lower than the actual
raising time, therefore affecting the start-up time of the interactive
application. This commit attempts to fix this issue.