tasks_vs_rw.sh: change address of Linux git repository
The tasks_vs_rw.sh script checks if a Linux git repository has
already been cloned, in case it hasn't it clones it anew. The
URL of the Linux official repository has been changed in 2012,
this commit adapts the script to use the new address. This
commit also uses the https version of the URL to make the script
compatible with operating environment which apply firewall rules
to outgoing/incoming connections.