video_play_vs_comms.sh: stop iterations after a timeout
This commit lets the main script stop the subsequent iterations if the
current takes too long to complete. The timeout is not implemented with
a killer thread as in comm_startup_lat.sh because, unlike that benchmark
script, video_play_vs_comms.sh invokes both mplayer and the interactive
processes as threads, in a non-blocking fashion.
The timeout is set according to the duration of the video playback and
takes into account a user-configurable tolerance.