video_play_vs_comms.sh: relax tolerance of the timeout
In the presence of a heavy write workload and BFQ, it may happen that
the video playback terminates correctly, but mplayer terminates only
after ten or more seconds elapse.
This sequence of events uncorrectly triggers the timeout which stops
subsequent iterations. This commit increases the tolerance of the
timeout in an attempt to avoid the issue