fairness.sh: terminate dd and iostat on sigint
Currently, the fairness.sh script uses the trap builtin to terminate
all related processes when the user forcedly terminates the test
with a Ctrl+C combination. However, the semantic of the shutdwn
function, used to perform the task, recently changed: now it wants
as a parameter a list of names of processes to be terminated. This
commit changes the usage of the shutdwn function accordingly.