utilities: first attempt at refactoring wr-duration-related code
Currently, some of the scripts in the benchmark suite rely on reading
BFQ's weight-raising period duration from a BFQ tunable as, from BFQ-v5
on, it is computed dynamically.
Starting from this piece of information, the scripts can compute a
correct amount of time to wait after starting a background workload,
so that e.g., an interactive application can be loaded when the
background processes' weight-raising period has already ended.
This commit attempts to factor out to a function the code computing
such an amount of time. Also, BFQ-v7r3 might introduce a change in
the tunable from which the current duration of a weight-raising period
can be read; this commit attempts to introduce support for both the
admittable file names.