kern_dev_tasks_vs_rw.sh: use a shorter interval for iostat
This commit changes the interval between iostat samples to be one
second; the change is instrumental to one of the following commits,
that will allow a command with a variable duration. In case that
a command's execution is very brief, the number of samples with
a large interval could not be sufficient. This commit mitigates
such a problem.