config_params.sh: move temp dir to /var/lib/bfq
The temp dir /tmp is just 7.8 GB on my system and mounted to
RAM. In other context I have seen /var/lib suggested as
something that can be relied on to always be local, even
for virtualized machines.

Add some tests to create the BASE_DIR if it doesn't exist
and fail and fall back to /tmp/test if it cannot be written.

The tests seem to be designed to be run as root so I guess
it may not be a major problem.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Acked-by: Paolo Valente <paolo.valente@unimore.it>