lib_utils: create and read separate files also for random I/O
A very large file was created in case of random I/O, instead of multiple
files as for sequential I/O. The reason was to increase the sector span
of I/O requests in case of random I/O. Unfortunately, on a non-rotational
device this can cause the following problem with several random readers:
because of the high speed of the device even with random I/O, each reader
may get to read a high percentage of the file soon enough, to let some
other reader find data in the page cachei during the test. This problem
does not occur with separate files, because fio guarantees that the cache
is invalidated for each file, once all the file has been read. For this
reason, this commit modifies random-file creation and reading, so that
multiple files are created, and then read separately.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>