check_dependencies: report all failed dependencies at once
check_dependencies stops, and aborts the whole benchmark, as it finds
a missing dependecy. This is annoying, because one then discovers the
next failed dependency only whne retrying to start the benchmark, and
so on. This commit fixes this problem by making check_dependencies
check all dependencies at once and report all failed ones.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>