bandwidth-latency: fix interfered fio pid to kill
FIO_PID variable is not the real interfered fio pid but the pid of the
subshell executing it in background. This can lead to misunderstandings.

Since the invocation command of the function start_fio_jobs is always
terminated with the control operator '&', the command which invokes this
function executes asynchronously in a subshell. For this reason the
start_fio_jobs function can not set variables in the parent shell (like
the global variable FIO_PID)

Thus, in order to share the interfered fio pid to kill with the function
execute_intfered_and_shutdwn_intferers, the start_fio_jobs function now
simply writes it to a temporary file, which can be read when needed.

Signed-off-by: Federico Motta <federico@willer.it>