bandwidth-latency: fix wait on interfered-stats.txt write completion
Synchronization between:
1) the end of start_fio_jobs function
2) and the start of compute_statistics function
is done with a sleep, which is not a solution for this kind of problem.

Since (1) is executing in a subshell which is not child of (2), (2) can
not wait for the end of (1). Nor (2) can wait for the end of the
subshell executing (1), since (1) issues writes which could not happen
instantly in the o.s.

Thus, this synchronization problem is solved with a "witness pass":
since (2) needs a file generated by (1), (1) also appends to that file a
"magic line" which (2) looks for, in order to be sure that (1) is ended.
The "magic line" / witness is also removed from the file in question at
the start of function (2).

Signed-off-by: Federico Motta <federico@willer.it>