file-copy: improve quality of test by adding extra features
In this test, many important features are missing.

In particular, there isn't any possibily to remove maxrate during file copy,
and there are no instructions to calculate execution time.

This commit adds the possibility to execute without a maxrate by replacing
the previous default value (16500 kB/s) with 0, that now implies an execution
without a maximum rate.
In this particular case (maxrate=0), it's also possible to set a new variable
"$PIPE" with the values "pipe" (default) or "nopipe", in order to execute the
two dd instructions respectively with and without a pipe.
It has also been added a time instruction, near the wait, that prints the
execution time of the dd instructions.